cart refactor
Former-commit-id: aa45964f6875da9131f884ea7d9683f5378a9e9d

+++ core/testing/sku_test.go
@@ -13,7 +13,7 @@ import (
 
 var (
 	modelId int32 = 5
-	itemId  int32 = 1
+	itemId  int64 = 1
 )
 
 // 测试商品模型设置品牌
